Message creation MVP
Very basic message creation. I do not think that the sender_id or
recipient_id should be part of the permitted  params (maybe the
recipient_id can?). The sender_id should be coming from a
`before_action` hook that verifies the current_user and sends it
that way. For now, to get the API working, I will commit this.