Set up rails associations
Since the user can be the receiver or sender, I wanted to specify two
different `has_many`s with each foreign key.

In order to get a full list of messages for a specific user, we need a
way to query the user's messages by checking both of those foreign key
fields. We can get back a list of all messages with User#messages.