Add fake authentication method to set the user
In my mind, the user gets authenticated here and is set to an ivar so
the API controller can use that current user. Whoever is making the
request, should be able to receive their messages (recent or whatever)
and be able to send messages to a recipient.

Since we set the sender to the current user, we can remove that from the
permitted params.