Show messages for user with specific sender
Using the show action here isn't quite restful, but it felt easier to do
it here for time purposes. If I were to give this a full run, I think I
could make a new controller (`RecentMessagesController`?) I'd have to
think about it a bit more. For this exercise, I believe this is nice and
concise and gets the point across.