Get requests specs green
Most of my specs are being run in the model specs. I've adjusted the
request specs here to ensure the routes are successful and render the
proper status codes. I do not believe we need to test out the `recent`
scope here again since we do that in the model. For extra coverage, we
definitely can add some robust model specs here, but I'd like to keep
moving.